feat: Adiciona teste de emissão de evento de clique
